Remove modes from theme, renaming generate theme to be clearer about what it actually does now
