--- packages/components/src/ui/styles/theme/dark-high-contrast-mode-config.js
@@ -1,1 +1,0 @@
-export const darkHighContrastModeConfig = {};
