--- packages/components/src/ui/styles/theme/dark-mode-config.js
@@ -1,7 +1,0 @@
-export const darkModeConfig = {
-	colorBodyBackground: '#18191A',
-	colorText: '#E4E6EB',
-	colorTextMuted: '#7a7a7a',
-	colorTextInverted: '#050505',
-	colorTextHeading: '#ffffff',
-};
