--- packages/components/src/ui/styles/theme/high-contrast-mode-config.js
@@ -1,7 +1,0 @@
-export const highContrastModeConfig = {
-	colorDivider: '#444',
-	controlBorderColor: '#444',
-	controlBorderColorHover: '#444',
-	controlBorderColorSubtle: '#444',
-	surfaceBorderColor: '#444',
-};
