+++ packages/components/src/ui/styles/theme/theme.js
@@ -1,4 +1,1 @@
 export { config } from './config';
-export { darkModeConfig } from './dark-mode-config';
-export { highContrastModeConfig } from './high-contrast-mode-config';
-export { darkHighContrastModeConfig } from './dark-high-contrast-mode-config';
